+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Fixes
+* Fixed ports not getting bound when running a container with an `EXPOSE` directive in the image Dockerfile
 
 # 0.2.2
 ### Fixes